fix(moysklad): decode fractional stockDays as float64
MoySklad's /report/stock/all returns stockDays as a fractional number
(e.g. 410.44), but StockRow.StockDays was typed int, so JSON decoding
failed with "cannot unmarshal number 410.44 into ... of type int". The
whole stock report failed to parse, breaking every stock-based tool
(dead_stock, get_stock) — the agent then improvised a "temporarily
unavailable" fallback whenever a user asked about остатки.

Type StockDays as float64 to match the wire; presentation still emits a
whole number of days via int(r.StockDays), and the dead-stock threshold
compares against float64(thresholdDays). Adds a decode regression test.

+++ internal/aggregate/analytics.go
@@ -283,7 +283,7 @@ type DeadStockLine struct {
 func DeadStock(rows []moysklad.StockRow, outcomeByRef map[string]float64, thresholdDays int) []DeadStockLine {
 	var out []DeadStockLine
 	for _, r := range rows {
-		if r.Stock <= 0 || r.StockDays < thresholdDays {
+		if r.Stock <= 0 || r.StockDays < float64(thresholdDays) {
 			continue
 		}
 		outcome := 0.0
